Add helper for fs.readFile()
